refactor : #185 CustomException 단일화 및 디렉터리 경로 정리
