+++ com.taken_seat.common-service/src/main/java/com/taken_seat/common_service/exception/customException/PaymentException.java
@@ -2,14 +2,14 @@ package com.taken_seat.common_service.exception.customException;
 
 import com.taken_seat.common_service.exception.enums.ResponseCode;
 
-public class PaymentNotFoundException extends BaseException {
+public class PaymentException extends BaseException {
 
-	public PaymentNotFoundException(ResponseCode responseCode) {
+	public PaymentException(ResponseCode responseCode) {
 		super(responseCode, responseCode.getMessage());
 
 	}
 
-	public PaymentNotFoundException(ResponseCode errorCode, String message) {
+	public PaymentException(ResponseCode errorCode, String message) {
 		super(errorCode, message);
 	}
 }
