--- com.taken_seat.common-service/src/main/java/com/taken_seat/common_service/exception/customException/PaymentHistoryNotFoundException.java
@@ -1,13 +1,0 @@
-package com.taken_seat.common_service.exception.customException;
-
-import com.taken_seat.common_service.exception.enums.ResponseCode;
-
-public class PaymentHistoryNotFoundException extends BaseException {
-	public PaymentHistoryNotFoundException(ResponseCode errorCode) {
-		super(errorCode);
-	}
-
-	public PaymentHistoryNotFoundException(ResponseCode errorCode, String message) {
-		super(errorCode, message);
-	}
-}
