+++ com.taken_seat.payment-service/src/main/java/com/taken_seat/payment_service/infrastructure/config/redis/RedisCachingConfig.java
@@ -1,4 +1,4 @@
-package com.taken_seat.payment_service.infrastructure.redis;
+package com.taken_seat.payment_service.infrastructure.config.redis;
 
 import java.time.Duration;
 import java.util.HashMap;
